backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
eventkit==1.0.3
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
ib-insync==0.9.86
idna==3.11
nest-asyncio==1.6.0
numpy==2.4.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
